+++ R/module.R
@@ -2,7 +2,7 @@
 #'
 #' @param paper a paper object or a list of paper objects
 #' @param module the name of a module or path to a module to run on this object
-#' @param ... further arguments to the module (e.g., arguments for the `llm()` function like `seed`)
+#' @param ... further arguments to the module (e.g., arguments for the `llm()` function like `seed`); these will override any arguments in the module
 #'
 #' @return a list of the returned table and report text
 #' @export
@@ -22,7 +22,7 @@ module_run <- function(paper, module, ...) {
   })
   module_dir <- dirname(module_path)
 
-  mod_chunks <- names(json) %in% c("text", "code", "ml", "llm") |>
+  mod_chunks <- names(json) %in% c("text", "code", "llm") |>
     which()
 
   results <- paper
@@ -32,14 +32,16 @@ module_run <- function(paper, module, ...) {
       results <- results$table
     }
 
-    args <- c(json[[chunk]], list(...))
+    args <- json[[chunk]]
+    override_args <- list(...)
+    args[names(override_args)] <- override_args
 
     if (type == "text") {
       results <- module_run_text(results, args)
     } else if (type == "code") {
       results <- module_run_code(results, args, module_dir)
-    } else if (type == "ml") {
-      results <- module_run_ml(results, args, module_dir)
+    # } else if (type == "ml") {
+    #   results <- module_run_ml(results, args, module_dir)
     } else if (type == "llm") {
       results <- module_run_llm(results, args)
     } else {
@@ -126,43 +128,43 @@ module_run_code <- function(paper, args, module_dir = ".") {
   return(results)
 }
 
-#' Run machine learning module
+#' #' Run machine learning module
+#' #'
+#' #' @param paper the paper object (or list of objects)
+#' #' @param args a list of arguments
+#' #' @param module_dir the base directory for the module, in case resources are in files with relative paths
+#' #'
+#' #' @return data frame
+#' #' @keywords internal
+#' module_run_ml <- function(paper, args, module_dir = ".") {
+#'   model_dir <- file.path(module_dir, args$model_dir)
+#'   if (!file.exists(model_dir)) {
+#'     stop("The model directory ", args$model_dir, " could not be found; make sure the module specification file is using a relative path to the directory")
+#'   }
 #'
-#' @param paper the paper object (or list of objects)
-#' @param args a list of arguments
-#' @param module_dir the base directory for the module, in case resources are in files with relative paths
+#'   if (is.vector(paper)) {
+#'     text <- data.frame(text = paper)
+#'   } else if (is.data.frame(paper)) {
+#'     text <- paper
+#'   } else {
+#'     text <- search_text(paper, return = "sentence")
+#'   }
 #'
-#' @return data frame
-#' @keywords internal
-module_run_ml <- function(paper, args, module_dir = ".") {
-  model_dir <- file.path(module_dir, args$model_dir)
-  if (!file.exists(model_dir)) {
-    stop("The model directory ", args$model_dir, " could not be found; make sure the module specification file is using a relative path to the directory")
-  }
-
-  if (is.vector(paper)) {
-    text <- data.frame(text = paper)
-  } else if (is.data.frame(paper)) {
-    text <- paper
-  } else {
-    text <- search_text(paper, return = "sentence")
-  }
-
-  class_col <- args$class_col %||% "class"
-  return_prob <- args$return_prob %||% FALSE
-
-  results <- ml(text, model_dir,
-     class_col = class_col,
-     map = args$map,
-     return_prob = return_prob)
-
-  if (!is.null(args$filter)) {
-    keep <- results[[class_col]] %in% args$filter
-    results <- results[keep, ]
-  }
-
-  return( list(table = results) )
-}
+#'   class_col <- args$class_col %||% "class"
+#'   return_prob <- args$return_prob %||% FALSE
+#'
+#'   results <- ml(text, model_dir,
+#'      class_col = class_col,
+#'      map = args$map,
+#'      return_prob = return_prob)
+#'
+#'   if (!is.null(args$filter)) {
+#'     keep <- results[[class_col]] %in% args$filter
+#'     results <- results[keep, ]
+#'   }
+#'
+#'   return( list(table = results) )
+#' }
 
 #' Run LLM module
 #'
@@ -174,6 +176,7 @@ module_run_ml <- function(paper, args, module_dir = ".") {
 module_run_llm <- function(paper, args) {
   args$text <- search_text(paper)
   results <- do.call(llm, args)
+  # attr(results, "llm")
 
   return( list(table = results) )
 }