+++ tests/testthat/test-01-utils.R
@@ -37,11 +37,26 @@ test_that("concat_tables", {
   expect_equal(length(ids), length(papers))
 })
 
+
+test_that("is_paper_list", {
+  expect_equal(is_paper_list(psychsci), TRUE)
+  expect_equal(is_paper_list(psychsci[1]), TRUE)
+  expect_equal(is_paper_list(psychsci[[1]]), FALSE)
+  expect_equal(is_paper_list(list(1,3,5)), FALSE)
+  expect_equal(is_paper_list(NULL), FALSE)
+
+  # empty lists return TRUE
+  expect_equal(is_paper_list(psychsci[c()]), TRUE)
+  expect_equal(is_paper_list(list()), TRUE)
+})
+
 test_that("print.scivrs_paper", {
   paper <- demoxml() |> read_grobid()
   op <- capture_output(print(paper))
   op.sv <- capture_output(print.scivrs_paper(paper))
+  expected <- "---------------\nto_err_is_human\n---------------\n\nTo Err is Human: An Empirical Investigation\n\n* Sections: 4\n* Sentences: 24\n* References: 2\n* Citations: 2\n"
 
+  expect_equal(op, expected)
   expect_equal(op, op.sv)
   expect_true(grepl("to_err_is_human", op))
 })
@@ -51,4 +66,12 @@ test_that("verbose", {
   expect_equal(verbose(), FALSE)
   expect_equal(verbose(TRUE), TRUE)
   expect_equal(verbose(), TRUE)
+  expect_equal(verbose(0), FALSE)
+  expect_equal(verbose("FALSE"), FALSE)
+  expect_equal(verbose(1), TRUE)
+  expect_equal(verbose("TRUE"), TRUE)
+
+  expect_error(verbose("G"))
+  expect_invisible(verbose(TRUE))
+  expect_visible(verbose())
 })