+++ tests/testthat/test-09-report.R
@@ -113,3 +113,29 @@ test_that("detected", {
   expect_true(file.exists(pdf))
   # browseURL(pdf)
 })
+
+test_that("module_report", {
+  expect_true(is.function(papercheck::module_report))
+
+  expect_error(module_report())
+
+  # set up module output
+  module_output <- module_run(psychsci[1:4], "all-p-values")
+
+  report <- module_report(module_output)
+  expect_true(grepl("Showing 81 of 81 rows", report))
+  expect_true(grepl("^## List All P-Values \\{\\.info\\}", report))
+
+  report <- module_report(module_output, header = 3, maxrows = 20, trunc_cell = 10)
+  expect_true(grepl("Showing 20 of 81 rows", report))
+  expect_true(grepl("^### List All P-Values \\{\\.info\\}", report))
+  expect_true(grepl("Eye mov...", report, fixed = TRUE))
+
+  report <- module_report(module_output, header = "Custom header")
+  expect_true(grepl("^Custom header", report))
+
+  op <- capture_output(print(module_output))
+  expect_true(grepl("^\\|text     \\|section \\|header ", op))
+  expect_true(grepl("\n\nShowing 20 of 81 rows$", op))
+  expect_true(grepl("p = \\.237 \\|results \\|Eye movement strategies", op))
+})